[CircleGraph] Relocate message handlers
This will relocate message handler related codes to CircleGraphCtrl.

ONE-vscode-DCO-1.0-Signed-off-by: SaeHie Park <saehie.park@gmail.com>
